refactor(twitch-wheel): renamed paths to fit new component names
