--- src/components/twitch-wheel/OptionsMenu/OptionsMenu.css
@@ -1,46 +1,0 @@
-
-#options-menu .accordion-body {
-    padding: calc(var(--bs-accordion-body-padding-y) / 3) calc(var(--bs-accordion-body-padding-x) / 2);
-}
-#options-menu .options-menu-items .btn {
-    display: block;
-    text-align: left !important;
-    font-size: calc(1.275rem + 0.3vw) !important;
-    font-weight: 500 !important;
-    opacity:0.8;
-    transition: opacity 0.125s linear;
-    padding-left: 0;
-    padding-right: 0;
-}
-#options-menu .options-menu-items .btn:hover {
-    opacity:1;
-}
-#options-menu .options-menu-items .btn:disabled {
-    opacity: 0.2;
-}
-#options-menu .options-menu-items .btn.subsetting::before {
-    content: '\2514';
-    margin-left: -5px;
-    margin-right: 1px;
-    /* content: '\231E'; */
-    display: inline-block;
-    color: #555;
-}
-/* #options-menu .options-menu-items .btn.subsetting:has(+ .btn.subsetting)::before {
-    content: '\251C';
-} */
-#options-menu .btn.btn-link,
-#options-menu .btn.btn-link:hover {
-    text-decoration: none;
-}
-#options-menu .btn-check:checked + .btn,
-#options-menu :not(.btn-check) + .btn:active,
-#options-menu .btn:first-child:active,
-#options-menu .btn.active, .btn.show {
-    color: #000;
-    background-color: #ffffffaa;
-}
-#dropdown-debug-menu-items.dropup .dropdown-toggle::after {
-    /* hides the dropdown caret */
-    content: unset;
-}
