--- src/components/twitch-wheel/PlayerSelect/PlayerSelect.css
@@ -1,248 +1,0 @@
-button.change-col {
-    text-transform: uppercase;
-    margin: 3px;
-    font-size: 11px;
-    background-color: lightgray;
-}
-
-div.change-col-buttons-container {
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-end;
-    visibility: hidden;
-    margin-top: -25px;
-}
-
-div.change-col-buttons-container:hover {
-    display: flex;
-    flex-direction: row;
-}
-
-div.change-col-buttons-container .change-col {
-    background-color: tomato;
-    border-radius: 1rem;
-}
-div.change-col-buttons-container .change-col.send-code {
-    background-color: #d3d3d3;
-}
-div.change-col-buttons-container .change-col.send-code:hover {
-    background-color: #e3e3e3;
-}
-
-.player-card:hover {
-    background-color: dimgray;
-}
-
-.player-card:hover > .change-col-buttons-container {
-    visibility: visible;
-}
-
-.player-count {
-    font-size: 16px;
-    margin-top: 5px;
-}
-
-.player-count.overlimit {
-    font-weight: bold;
-    color: darkorange;
-}
-
-#main-screen button.btn.start-game {
-    padding-top: 0.5rem;
-    padding-bottom: 0.5rem;
-}
-#main-screen .card-header-item > .input-group + button.start-game {
-    padding-top: 1px;
-    padding-bottom: 1px;
-}
-
-button.start-game,
-.card-header-item > .input-group {
-    width: 100%;
-    margin-bottom: 5px;
-}
-
-input.room-code::placeholder {
-    /* letter-spacing: -0.5px; */
-    font-size: 90%;
-}
-
-.player-select .player-card-container {
-    overflow-y: auto;
-    height: calc(100vh - 250px);
-    font-size: 16px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.player-card-column {
-    width: 50%;
-    height: 100%;
-    margin: 0 5px;
-    border-radius: 5px;
-}
-
-.player-card-column.interested {
-    background-color: slategrey;
-}
-
-.player-card-column.playing {
-    background-color: teal;
-}
-
-.player-card-column.joined {
-    background-color: seagreen;
-    display: none;
-}
-
-.player-card-column > p {
-    margin-top: 0;
-    padding-top: 10px;
-    margin-bottom: 10px;
-    font-size: 17px;
-    font-weight: bold;
-}
-
-.player-card {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    background-color: gray;
-    margin: 8px;
-    border-radius: 1rem;
-    border-style: outset;
-    border-color: cadetblue;
-}
-
-.player-card-username {
-    display: flex;
-}
-
-.player-card-username > img {
-    width: 12px;
-    margin-left: 2px;
-}
-
-.player-name {
-    margin: 0px 0px 0px 16px;
-    font-weight: 500;
-    word-wrap: break-word;
-    text-align: left;
-}
-
-img + p.player-name {
-    margin-left: 0px;
-}
-
-button.dice {
-    width: 25px;
-    height: 25px;
-    background-color: greenyellow;
-    border-radius: 5px;
-    margin-left: 10px;
-    padding: 0px;
-    vertical-align: middle;
-    transform: scale(1) rotate(0deg);
-    transition: transform 0.2s;
-    cursor: pointer;
-}
-
-button.dice > img {
-    width: 100%;
-    height: 100%;
-    vertical-align: unset;
-}
-
-.toggle-streamer-seat,
-.card-header-item {
-    display: flex;
-    align-items: center;
-}
-
-.toggle-streamer-seat .toggle-label,
-.card-header-item-label {
-    font-size: 12px;
-    max-width: 40px;
-    overflow-wrap: break-word;
-    text-align: center;
-}
-.toggle-streamer-seat .toggle-label {
-    max-width: 80px;
-    text-align: right;
-}
-
-.toggle-streamer-seat .form-switch {
-    margin-left: 8px;
-    padding-left: 3em;
-}
-.toggle-streamer-seat .form-switch .form-check-input {
-    height: 1.6em;
-    width: 3em;
-    margin-left: -3em;
-    background-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' viewBox='-4 -4 8 8'%3e%3ccircle r='3' fill='%23fff'/%3e%3c/svg%3e");
-    background-position: left center;
-    border-radius: 3em;
-    transition: background-position 0.15s ease-in-out;
-    background-color: #f87171;
-    border-color: #f87171;
-}
-.toggle-streamer-seat .form-switch .form-check-input:focus {
-    background-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' viewBox='-4 -4 8 8'%3e%3ccircle r='3' stroke='%2386b7fe' stroke-width='0.5' fill='%23fdd'/%3e%3c/svg%3e");
-}
-.toggle-streamer-seat .form-check-input:checked {
-    background-color: #34d399;
-    border-color: #34d399;
-}
-input#reserve-seat-for-streamer:after {
-    display: block;
-    content: "no";
-    position: absolute;
-    color: #fff;
-    font-size: 0.7rem;
-    text-align: center;
-    width: 2.25em;
-    line-height: 2em;
-    font-weight: bold;
-    margin-left: 1.75em;
-}
-input#reserve-seat-for-streamer:checked:after {
-    content: "yes";
-    margin-left: 0;
-}
-
-.card.rand-1 .interested .player-card:nth-last-child(12n+2),
-.card.rand-2 .interested .player-card:nth-child(9n-3),
-.card.rand-3 .interested .player-card:nth-child(4n-3),
-.card.rand-4 .interested .player-card:nth-child(3n+2),
-.card.rand-5 .interested .player-card:nth-child(5n+4),
-.card.rand-6 .interested .player-card:nth-child(6n-4),
-.card.rand-7 .interested .player-card:nth-child(3n+3),
-.card.rand-8 .interested .player-card:nth-child(4n+1),
-.card.rand-9 .interested .player-card:nth-child(4n-2),
-.card.rand-10 .interested .player-card:nth-child(8n-3),
-.card.rand-11 .interested .player-card:nth-child(6n-2),
-.card.rand-12 .interested .player-card:nth-child(4n+1),
-.card.rand-13 .interested .player-card:nth-child(5n-3),
-.card.rand-14 .interested .player-card:nth-last-child(11n+3) {
-    background-color: hsl(0deg 0% 80% / 75%);
-}
-.card:not(.rand-0) button.dice {
-    transition: all 0.15s;
-    animation: tilt-shaking 0.15s infinite;
-    pointer-events: none;
-}
-.card:not(.rand-0) button.dice img {
-    transform: scale(2);
-    border-radius: 5px;
-    background-color: greenyellow;
-    transition: all 0.15s;
-}
-
-@keyframes tilt-shaking {
-    0% { transform: scale(2) rotate(0deg); }
-    25% { transform: scale(2) rotate(5deg); }
-    50% { transform: scale(2) rotate(0deg); }
-    75% { transform: scale(2) rotate(-5deg); }
-    100% { transform: scale(2) rotate(0deg); }
-  }
